💇 Refactor build
Preliminary separations to facilitate conversion to ts script
