+++ package.json
@@ -9,7 +9,7 @@
     "start": "next start",
     "lint": "next lint",
     "update:submodules": "git submodule update --remote",
-    "pre-build": "node scripts/build.js"
+    "pre-build": "ts-node scripts/build.js"
   },
   "dependencies": {
     "gray-matter": "^4.0.3",
