+++ utils/createSlug.js
@@ -1,0 +1,5 @@
+function createSlug(fileName, frontmatter) {
+  return frontmatter?.slug ?? fileName.replace(".md", "");
+}
+
+exports.createSlug = createSlug;
